dodano kolejna biblioteke - kompatybilnosc z xamarinem
