+++ D2DLibrary/D2DLibrary.csproj
@@ -50,7 +50,6 @@
       <HintPath>..\packages\System.Buffers.4.4.0\lib\netstandard2.0\System.Buffers.dll</HintPath>
     </Reference>
     <Reference Include="System.Core" />
-    <Reference Include="System.Drawing" />
     <Reference Include="System.Memory, Version=4.0.1.1, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL">
       <HintPath>..\packages\System.Memory.4.5.3\lib\netstandard2.0\System.Memory.dll</HintPath>
     </Reference>
@@ -71,10 +70,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AutoConfig.cs" />
-    <Compile Include="BMsending.cs" />
     <Compile Include="CommClient.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ScreenCapture.cs" />
     <Compile Include="Test.cs" />
   </ItemGroup>
   <ItemGroup>
