+++ D2DLibrary/Test.cs
@@ -1,9 +1,7 @@
-﻿using D2DLibrary;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Drawing;
-using System.Drawing.Imaging;
 using System.Linq;
 using System.Net;
 using System.Net.Sockets;
@@ -97,97 +95,6 @@ namespace TCPSender
             Console.ReadLine();
 
         }
-
-        public static void LZ4_Test()
-        {
-            CompressScreen display = new CompressScreen();
-            display.Iterate();
-
-
-            DecompressScreen dcmpDisplay = new DecompressScreen(display.getData());
-            int i = 0;
-            Bitmap test1 = dcmpDisplay.DAiB(dcmpDisplay.decompressed);
-
-            Console.ReadLine();
-        }
-
-        public static void LZ4_Test_DX()
-        {
-            CompressScreen display = new CompressScreen();
-            display.IterateDX();
-
-
-            DecompressScreen dcmpDisplay = new DecompressScreen(display.getData());
-            int i = 0;
-            Bitmap test1 = dcmpDisplay.DAiB(dcmpDisplay.decompressed);
-
-            Console.ReadLine();
-        }
-
-        public static void LZ4_PerformanceTest()
-        {
-            CompressScreen display = new CompressScreen();
-
-            for(int i = 0; i<100; i++)
-            {
-                display.PerformanceTest();
-            }
-
-            Console.ReadLine();
-        }
-
-        public static void LZ4_DX_PerformanceTest()
-        {
-            CompressScreen display = new CompressScreen();
-
-            for (int i = 0; i < 100; i++)
-            {
-                display.PerformanceTestDX();
-            }
-
-            Console.ReadLine();
-        }
-
-        public static void CaptureScreenTest()
-        {
-            Stopwatch sw = Stopwatch.StartNew();
-
-            Bitmap bitmapa = ScreenCapture.CaptureScreen();
-
-            TimeSpan timeToScreenCapture = sw.Elapsed;
-
-            Console.WriteLine("Screen: {0}ms", timeToScreenCapture.TotalMilliseconds);
-
-            Console.ReadLine();
-        }
-
-        public static void DXCaptureScreenTest()
-        {
-
-
-            List<double> lista = new List<double>();
-
-            var screenStateLogger = new ScreenStateLogger();
-            screenStateLogger.ScreenRefreshed += (sender, data) =>
-            {
-                
-            };
-            screenStateLogger.Start();
-            Thread.Sleep(1000);
-            screenStateLogger.Stop();
-
-            //for(int i=0; i < 10; i++)
-            //{
-            //    Console.WriteLine(lista[i]);
-            //}
-
-            //TimeSpan timeToScreenCapture = sw.Elapsed;
-
-            //Console.WriteLine("Screen: {0}ms", timeToScreenCapture.TotalMilliseconds);
-
-            Console.ReadLine();
-        }
-
 
     }
 }
