+++ ScreenCaptureLibrary/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="lz4net" version="1.0.15.93" targetFramework="net461" />
+  <package id="SharpDX" version="4.2.0" targetFramework="net461" />
+  <package id="SharpDX.Direct3D11" version="4.2.0" targetFramework="net461" />
+  <package id="SharpDX.DXGI" version="4.2.0" targetFramework="net461" />
+</packages>
