+++ TCPSender.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "D2DLibrary", "D2DLibrary\D2
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "D2DUIv3", "D2DUIv3\D2DUIv3.csproj", "{1F6ECA8E-AAA3-4EB1-BFA5-7EC855563130}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScreenCaptureLibrary", "ScreenCaptureLibrary\ScreenCaptureLibrary.csproj", "{6437B61C-93B1-46DE-8CC4-C5BCC4EF484E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -29,6 +31,10 @@ Global
 		{1F6ECA8E-AAA3-4EB1-BFA5-7EC855563130}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1F6ECA8E-AAA3-4EB1-BFA5-7EC855563130}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1F6ECA8E-AAA3-4EB1-BFA5-7EC855563130}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{6437B61C-93B1-46DE-8CC4-C5BCC4EF484E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6437B61C-93B1-46DE-8CC4-C5BCC4EF484E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6437B61C-93B1-46DE-8CC4-C5BCC4EF484E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6437B61C-93B1-46DE-8CC4-C5BCC4EF484E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
